Transform ElementFinderInstigator into an interface `TypedInstigator`
